res added functions to print formatted strings to stderr. res eliminated a static var from win32kbd.cpp. res silenced some MSVC warnings in regcomp.c and snprintf.cpp. res replaced some ints with size_ts in csPoly3D. res removed the csextern.h #inclusion from csutil/hashr.h. Not needed   there.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19050 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csgeom/math3d.cpp
@@ -451,7 +451,7 @@ bool csIntersect3::IntersectPolygon (
 {
   csVector3 &v1 = seg.Start ();
   csVector3 &v2 = seg.End ();
-  int i, i1;
+  size_t i, i1;
   float c, c1;
   csVector3 isect;
   float dist;
